ASMS integration, Addressed review comment pr-36 auto clock update, Added fix for  backspase key issue
