+++ src/components/CurrentTime/constants.js
@@ -1,0 +1,5 @@
+module.exports = {
+    TIME_LEFT: 1650,
+    TIME_TOP: 15,
+    TIME_ZINDEX: 11
+  }
